Refactor model processing and add InternalModel type hint
